chore: extract Provider piece logic to separate file
